3.MileStone1-Arrays.dcox SOLUTIONS
files contain:-
Solution.java -1.Write a program to initialize an integer array and print the sum and average of the array  
Solution2.java-2.Write a program to initialize an integer array and find the maximum and minimum value of an array.
Solution3.java-3.Write a program to initialize an integer array with values and check if a given number is present in the array or not. If the number is not found, it will print -1 else it will print the index value of the given  number in the array 
Solution4.java-4.Initialize an integer array with ascii values and print the corresponding character values in a single row.